foodRunServer/requirements.txt
Flask
flask-cors
elevenlabs
python-dotenv
aws-wsgi
requests
boto3
spacy
sentence_transformers
word2number